LayoutTests/fast/parser/script-tag-with-trailing-slash.html

<html>

<script>
if (window.layoutTestController)
    layoutTestController.dumpAsText();
</script>

<body>
You should see the text "inside" the following script tag:
<script language="JavaScript" src=""/>
This is the text inside the script tag.
</body>

</html>	
