LayoutTests/fast/parser/script-tag-with-trailing-slash-expected.txt

CONSOLE MESSAGE: line 1: SyntaxError: Parse error
You should not see the text "inside" the following script tag: This is the text inside the script tag.
